added hover on habitsSideBar
