+++ src/components/SideBarHabits/SideBarHabits.module.css
@@ -45,3 +45,7 @@ table {
   align-items: center;
   border-bottom: 1px solid rgba(238, 240, 247, 0.1);
 }
+
+.iconsWrap {
+  margin-left: 130px;
+}
